ref(*): Refactors the history action to return releases instead of formatted output
Output utilities have been moved to their own file and streamlined as well
